webhook-nginx/requirements.txt
anyio==4.3.0
async-timeout==4.0.3
blinker==1.7.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
exceptiongroup==1.2.0
h11==0.14.0
httpcore==1.0.4
httpx==0.27.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
python-telegram-bot==21.0.1
redis==5.0.1
requests==2.31.0
sniffio==1.3.1
tornado==6.4
typing_extensions==4.10.0
tzlocal==5.2
urllib3==1.26.15
Werkzeug==3.0.1
